Overhaul static method api
Instead of request, resource, copts being passed in, a call site
context is passed as the last OPTIONAL argument. This arguments will
contain all information that can be decuded from the context / call site.

This greatly reduces the amount of code that needs to be written for static
methods and what is required or imported into each one. It acts as a sort of
call site middleware that can eval the method and deduce the resource uri
before passing the contexting to the static method.