fix postReject bug, adjust unit-tests
postReject may encounter an error without a response object, in this case
the problem was that this scenario was not handled and the postReject would throw.
update fixes this issue by forwarding the error when there is no response object
in the error passed.

updated and fixed unit-tests to pass